+++ CHANGELOG.md
@@ -2,3 +2,4 @@
     > **Note** : This is an unstable release
     - feat(web): Add new About page (#258) by @prayag17
     - feat(web): Allow managing servers from settings dialog by @prayag17
+    - feat(search): Add Episode to search by @prayag17